Feat : Migrated ServiceWorker to next-pwa/Workbox for Automated Precaching#3441
