+++ apps/web/next.config.mjs
@@ -1,7 +1,14 @@
 import { execSync } from "node:child_process";
 import createNextIntlPlugin from "next-intl/plugin";
+import withPWAInit from "@ducanh2912/next-pwa";
 
 const withNextIntl = createNextIntlPlugin();
+
+const withPWA = withPWAInit({
+    dest: "public",
+    disable: process.env.NODE_ENV === "development",
+});
+
 
 /**
  * Deterministic build ID derived from the Git commit SHA.
@@ -55,4 +62,4 @@ const nextConfig = {
     },
 };
 
-export default withNextIntl(nextConfig);
+export default withPWA(withNextIntl(nextConfig));
