+++ apps/web/package.json
@@ -3,9 +3,7 @@
     "version": "0.1.0",
     "private": true,
     "scripts": {
-        "predev": "node scripts/generate-sw.mjs",
         "dev": "next dev",
-        "prebuild": "node scripts/generate-sw.mjs",
         "build": "next build",
         "start": "next start",
         "lint": "eslint .",
@@ -15,8 +13,8 @@
         "format:check": "prettier --check ."
     },
     "dependencies": {
+        "@ducanh2912/next-pwa": "^10.2.9",
         "@google/genai": "^2.10.0",
-        "@sahidawa/validators": "*",
         "@hookform/resolvers": "^5.4.0",
         "@opentelemetry/auto-instrumentations-web": "^0.64.0",
         "@opentelemetry/context-zone": "^2.8.0",
@@ -24,6 +22,7 @@
         "@radix-ui/react-tooltip": "^1.2.11",
         "@sahidawa/shared": "*",
         "@sahidawa/types": "*",
+        "@sahidawa/validators": "*",
         "@supabase/auth-helpers-nextjs": "^0.15.0",
         "@supabase/ssr": "^0.12.0",
         "@supabase/supabase-js": "^2.110.0",
